Arraylist 관련 예제 커밋

+++ Practice/02_java/chap11-io-lecture-source/src/main/java/com/ohgiraffers/section03/filterstream/Application4.java
@@ -7,7 +7,7 @@ import java.io.*;
 public class Application4 {
     public static void main(String[] args) {
 
-        MemberDTO[] memArr = new MemberDTO[100];
+        MemberDTO[] memArr = new MemberDTO[50];
         memArr[0] = new MemberDTO("user01", "pass01", "홍길동",
                 "hong123@gmail.com", 25, '남');
         memArr[1] = new MemberDTO("user02", "pass02", "유관순",
@@ -15,18 +15,29 @@ public class Application4 {
         memArr[2] = new MemberDTO("user03", "pass03", "강감찬",
                 "kang@gmail.com", 38, '남');
 
-
+        File ObjFile = new File("src/main/java/com/ohgiraffers/section03/filterstream/testObject.txt");
         ObjectOutputStream oos = null;
         try {
-//            oos = new ObjectOutputStream(
-            oos = new MyOutput(
-                    new FileOutputStream(
-                            "src/main/java/com/ohgiraffers/section03/filterstream/testObject.txt"
-                    )
-            );
+            if(!ObjFile.exists()){
+                oos = new ObjectOutputStream(
+                        new BufferedOutputStream(
+                                new FileOutputStream(
+                                        "src/main/java/com/ohgiraffers/section03/filterstream/testObject.txt"
+                                )
+                        )
+                );
+            } else {
+                oos = new MyOutput(
+                        new BufferedOutputStream(
+                                new FileOutputStream(
+                                        "src/main/java/com/ohgiraffers/section03/filterstream/testObject.txt", true
+                                )
+                        )
+                );
+            }
 
 //            oos.writeObject(memArr[0]);
-            for (int i = 0; i < memArr.length; i++) {
+            for (int i = 0; i < 3; i++) {           // 온전한 객체만 들어가도록 실체가 있는 수만큼 반복
                 oos.writeObject(memArr[i]);
             }
 
@@ -34,9 +45,7 @@ public class Application4 {
             throw new RuntimeException(e);
         } finally {
             try {
-                if(oos != null) {
-                    oos.close();
-                }
+                if(oos != null) oos.close();
             } catch (IOException e) {
                 throw new RuntimeException(e);
             }
@@ -47,16 +56,20 @@ public class Application4 {
         ObjectInputStream ois = null;
         try {
             ois = new ObjectInputStream(
-                    new FileInputStream(
-                            "src/main/java/com/ohgiraffers/section03/filterstream/testObject.txt"
+                    new BufferedInputStream(
+                            new FileInputStream(
+                                    "src/main/java/com/ohgiraffers/section03/filterstream/testObject.txt"
+                            )
                     )
             );
 
 //            newMemArr[0] = (MemberDTO)ois.readObject();
             int index = 0;
-            while (true) {
+            while(true) {
+                if(index == newMemArr.length) break;
                 newMemArr[index++] = (MemberDTO) ois.readObject();
             }
+
         } catch (EOFException e) {
             System.out.println("회원 정보 읽기 완료!");
         } catch (IOException e) {
@@ -65,16 +78,15 @@ public class Application4 {
             throw new RuntimeException(e);
         } finally {
             try {
-                if(ois != null) {
-                    ois.close();
-                }
+                if(ois != null) ois.close();
             } catch (IOException e) {
                 throw new RuntimeException(e);
             }
         }
 
-        /* 설명. 출력 및 입력까지 잘 완료되었는지 새로운 배열의 회원 정보 확인*/
-        for(MemberDTO mem : newMemArr) {
+        /* 설명. 출력 및 입력까지 잘 완료되었는지 새로운 배열의 회원 정보 확인 */
+        for (MemberDTO mem : newMemArr) {
+            if(mem == null) break;
             System.out.println(mem);
         }
     }
